move things to plugins / transformers
