+++ packages/webawesome/docs/_plugins/replace-text.js
