+++ packages/webawesome/docs/_plugins/search.js
