+++ packages/webawesome/docs/_utils/simulate-webawesome-app.js
@@ -1,5 +1,8 @@
 import nunjucks from 'nunjucks';
 
+/**
+ * This function simulates what a server would do running "on top" of eleventy.
+ */
 export function SimulateWebAwesomeApp(str) {
   return nunjucks.renderString(str, {
     // Stub the server EJS shortcodes.
